27-12-2022

+++ auth/signup.php
@@ -1,27 +1,42 @@
 <?php
 include "../connect.php";
+
+$signUpLang = postData("signUpLang");
 $username = postData("username");
 $email = postData("email");
 $password = postData("password");
 
+//count_chars()
 
-$statement = $connect->prepare("INSERT INTO `users`( `username`, `email`, `password`) VALUES ('$username','$email','$password')");
+/// Check If there email in db
+$checkExistingEmail = $connect->prepare("SELECT * FROM `users` WHERE  email='$email'");
+// run last quiry
+$checkExistingEmail->execute();
 
-$statement->execute();
+$dataEmail = $checkExistingEmail->fetchAll(PDO::FETCH_ASSOC);
 
-$signupCount = $statement->rowCount();
+$isEmailExiste = count($dataEmail);
+
+if ($isEmailExiste == 0) {
 
 
-if ($signupCount > 0) {
+    $signUpStmt = $connect->prepare("INSERT INTO `users`( `username`, `email`, `password`) VALUES ('$username','$email','$password')");
 
-    echo json_encode(
-        array("state" => "success",
-            "add_user" => $username,
-            "add_email" => "$email",
-            "set_password" => "success",
-        ));
+    $signUpStmt->execute();
+
+    $signupCount = $signUpStmt->rowCount();
+
+
+    if ($signupCount > 0) {
+
+        echo json_encode(
+            langSuccess($signUpLang, $usernamem, $email),  );
+    } else {
+
+        echo json_encode(array("state" => "fail to add $username"));
+
+    }
 } else {
 
-    echo json_encode(array("state" => "fail to add $username"));
-
+    echo json_encode(langFail($signUpLang));
 }
